requirements.txt
solana==0.30.2
solders==0.18.1
base58==2.1.1
websockets==11.0.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
typing-extensions==4.8.0
logging==0.4.9.6
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
flake8==6.1.0
black==23.11.0
mypy==1.7.1
sphinx==7.2.6
sphinx-rtd-theme==1.3.0